more AbstractTile tests
